+++ semcore/illustration/CHANGELOG.md
@@ -2,7 +2,14 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [1.0.0-0] - unreleased
+## [1.0.0] - unrelease
+
+### Changed
+
+- Added react component `MailSent` put on path ``@semcore/illustration/MailSent`
+- Illustration `MailSent` put on path `@semcore/illustration/svg/MailSent`
+
+## [1.0.0-0] - 2022-02-28
 
 ### Added